#379 [FIX] empty_view
정보 등록글 없을 때 이상해지는 부분 수정
