+++ app/src/main/res/layout/fragment_class_room.xml
@@ -13,6 +13,39 @@
         android:layout_height="match_parent"
         tools:context=".app.presentation.ui.classroom.ClassRoomFragment">
 
+        <TextView
+            android:id="@+id/text_info_empty"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/information_no_item"
+            android:textColor="@color/gray_3"
+            android:fontFamily="@font/pretendard_regular"
+            android:textSize="14sp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"/>
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btn_go_information_write"
+            android:layout_width="64dp"
+            android:layout_height="0dp"
+            android:layout_marginEnd="24dp"
+            android:layout_marginBottom="24dp"
+            android:background="@drawable/oval_fill_black"
+            android:drawableTop="@drawable/ic_plus"
+            android:elevation="25dp"
+            android:fontFamily="@font/pretendard_regular"
+            android:paddingTop="14dp"
+            android:paddingBottom="10dp"
+            android:text="@string/classroom_info_title"
+            android:textColor="@color/main_light"
+            android:textSize="11sp"
+            app:backgroundTint="@color/main_black"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintDimensionRatio="1"
+            app:layout_constraintEnd_toEndOf="parent" />
+
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/cl_class_sort"
             android:layout_width="match_parent"
@@ -53,10 +86,11 @@
             app:layout_constraintTop_toBottomOf="@+id/cl_class_sort"
             app:layout_constraintBottom_toBottomOf="parent">
 
+
+
             <androidx.core.widget.NestedScrollView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:overScrollMode="never"
                 android:fillViewport="true">
 
                 <androidx.constraintlayout.widget.ConstraintLayout
@@ -101,10 +135,10 @@
                     <androidx.fragment.app.FragmentContainerView
                         android:id="@+id/fragment_container_classroom"
                         android:layout_width="match_parent"
-                        android:layout_height="0dp"
+                        android:layout_height="wrap_content"
                         android:layout_marginTop="16dp"
-                        app:layout_constraintTop_toBottomOf="@+id/text_classroom_info_title"
-                        app:layout_constraintBottom_toBottomOf="parent"/>
+                        android:layout_marginBottom="20dp"
+                        app:layout_constraintTop_toBottomOf="@+id/text_classroom_info_title" />
 
 
                 </androidx.constraintlayout.widget.ConstraintLayout>
