+++ app/src/main/res/layout/fragment_information.xml
@@ -12,39 +12,7 @@
         android:layout_height="match_parent"
         tools:context=".app.presentation.ui.classroom.InformationFragment">
 
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/btn_go_information_write"
-            android:layout_width="64dp"
-            android:layout_height="0dp"
-            android:layout_marginEnd="24dp"
-            android:layout_marginBottom="24dp"
-            android:background="@drawable/oval_fill_black"
-            android:drawableTop="@drawable/ic_plus"
-            android:elevation="25dp"
-            android:fontFamily="@font/pretendard_regular"
-            android:paddingTop="14dp"
-            android:paddingBottom="10dp"
-            android:text="@string/classroom_info_title"
-            android:textColor="@color/main_light"
-            android:textSize="11sp"
-            app:backgroundTint="@color/main_black"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintDimensionRatio="1"
-            app:layout_constraintEnd_toEndOf="parent" />
 
-
-        <TextView
-            android:id="@+id/text_info_empty"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/information_no_item"
-            android:textColor="@color/gray_3"
-            android:fontFamily="@font/pretendard_regular"
-            android:textSize="14sp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"/>
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
